EPMRPP-112184-extract-sso || Implement authentication extension and refactor auth integration logic
